白敵の処理を修正完了。立ち絵追加。現存で気になるバグは所持クナイ数に表示が対応してないこと

+++ AnimationComponent.h
@@ -44,5 +44,5 @@ private:
     std::string m_currentName;
 
     float m_timer;
-    float m_frameIndex;
+    int m_frameIndex;
 };
